+++ Common/src/main/java/gay/object/hexdebug/mixin/HexDebugMixinConfigPlugin.java
@@ -1,5 +1,6 @@
 package gay.object.hexdebug.mixin;
 
+import dev.architectury.platform.Platform;
 import gay.object.hexdebug.HexDebug;
 import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
@@ -26,6 +27,10 @@ public class HexDebugMixinConfigPlugin implements IMixinConfigPlugin {
                 HexDebug.LOGGER.warn("Applying scuffed datagen mixin. This should not happen if not running datagen!");
             }
             return shouldApply;
+        }
+        if (mixinClassName.startsWith("gay.object.hexdebug.mixin.interop.")) {
+            var id = mixinClassName.substring("gay.object.hexdebug.mixin.interop.".length()).split("\\.", 2)[0];
+            return Platform.isModLoaded(id);
         }
         return true;
     }